+++ config/theme.js
@@ -62,5 +62,9 @@ export default {
 
   divider: {
     padding: 16
+  },
+  formItem: {
+    padding: 24,
+    labelWidth: 140
   }
 }